Create and move gui to subpackage
